feat: adjust header design
