+++ src/app/admin/layout.tsx
@@ -3,7 +3,6 @@
 import {observer} from 'mobx-react-lite'
 import {ReactNode} from 'react'
 
-import {authStore} from '../auth/context/auth-store'
 import {Header} from './components'
 import {SmartphoneProvider} from './context/smartphone-context'
 import {useValidateAuth} from './hooks'
@@ -19,7 +18,7 @@ const AdminLayout = observer(({children}: AdminLayoutProps) => {
     <SmartphoneProvider>
       <div className='flex h-screen w-screen flex-col items-center overflow-auto bg-gray-900 p-3'>
         <div className=' mb-3 w-full '>
-          <Header user={authStore.user} />
+          <Header />
         </div>
         {children}
       </div>
